Spain Vioscreen Bypass
For THDMI Spain, we made the decision to hide the Vioscreen option to minimize confusion, as it won't work in Spain. This behavior is working correctly on the Source page, as the Vioscreen FFQ button is not displayed for Spanish users. However, during testing, we missed the post-sample page that encourages the user to take the Vioscreen FFQ.

This change mirrors the logic of determining a Spanish user (either their account country or locale matching Spain) and, if it does, bypasses the redirect to the page suggesting they take the Vioscreen FFQ.