Update pip and conda requirements to fix error when running server after repo install
After installing the code in the repo according to steps in the README.md file, an error
occurs when running the server with the command $ python microsetta_interface/server.py

Error: ImportError: cannot import name 'final' from 'typing'
(/usr/local/anaconda3/envs/microsetta-interface_test/lib/python3.7/typing.py)
The error encountered occurs because the final feature was introduced in Python 3.8 and
is not available in Python 3.7. The error occurs in async_timeout which
depends on redis. The version of redis installed is 5.0.7. On the pypi page for redis,
at https://pypi.org/project/redis/, this statement is found:
** Note: ** redis-py 5.0 will be the last version of redis-py to support Python 3.7,
as it has reached end of life. redis-py 5.1 will support Python 3.8+.
Uninstalling redis and installing redis==4.6.0 fixes the error when starting the server.

Used pip list and conda list to update the versions in pip_requirements.txt and
conda_requirements.txt. Keeping the same packages in each requirement file with the
updated versions resulted in an error with a fresh install and running the server.

To fix the error, removed redis==4.6.0 from conda_requirements.txt because conda could
not find this version of redis and put it in the pip_requirements.txt file. Also needed
to add babel as a package to conda_requirements.xt since there was a
ModuleNotFoundError for babel. Update flask-babel=2.0.0 in conda_requirements.txt to
flask-babel==2.0.0 replace a single = to a double =. Remove pytest==5.3.3 from
conda_requirements.txt since it is already in pip_requirements.txt.