Modify pip and conda requirements files
Pin only the redis package. Set the version to 4.6.0 in pip_requirements.txt to fix error
when running the server after an install

Remove redis from conda_requirements.txt and place in pip_requirements.txt because conda
can't locate the redis version 4.6.0.

Remove pytest from conda_requirements.txt because it was not versioned and there is a
versioned pytest < 5.3.4 in pip_requirements.txt.

Add an equal sign to "flask-babel=2.0.0" in conda_requirements.txt to set the version
using double equal signs.