Update pip and conda requirements for python server to work in conda environ
Remove redis-py and redis from coda requirements and add redis
to pip_requirements.txt. It's not possible to install redis-server with
pip or conda since it's not a python package. So the only install for
redis is the client. Pip has a more current redis version than conda.
With conda, redis needs to be < 5.0 and for pip, redis is 5.0.8. Having
redis only in pip_requirements also follows the setup for
microsetta-private-api where redis is in pip_requirements.txt and
not in conda_requirements.txt.