Remove radial score of sample diversity and update image size calculation
The sample diversity score is not placed next to the dial and the function
recalculating gauge distances is simplified by removing any code adjusting
the sample diversity radial score.

The diversity gauge image width is adjusted to respond to the width of
descriptive text in different languages.

Remove hard coded sample diversity score and get it from the ajax call
as it was originally done.